+++ item_test.go
@@ -252,3 +252,73 @@ func TestService_UpdateItem(t *testing.T) {
 		})
 	}
 }
+
+func TestService_DeleteItem(t *testing.T) {
+	tests := []struct {
+		name     string
+		uuid     uuid.UUID
+		exchange *microtest.Exchange
+		uri      string
+		e        error
+	}{
+		{
+			name: "403 permission required",
+			uuid: uuid.MustParse("40355dba-0923-43a6-83d5-c9b6680edd2e"),
+			exchange: &microtest.Exchange{
+				Response: microtest.Response{
+					Status: 403,
+					Body:   noPermission,
+				},
+			},
+			uri: "/item/40355dba-0923-43a6-83d5-c9b6680edd2e",
+			e:   errors.New("no permission"),
+		},
+		{
+			name: "404 item not found",
+			uuid: uuid.MustParse("f27ef50d-f10f-4ff8-b65a-d64b1ebb83c8"),
+			exchange: &microtest.Exchange{
+				Response: microtest.Response{
+					Status: 404,
+					Body:   notFound,
+				},
+			},
+			uri: "/item/f27ef50d-f10f-4ff8-b65a-d64b1ebb83c8",
+			e:   errors.New("not found"),
+		},
+		{
+			name: "200 successful",
+			uuid: uuid.MustParse("f27ef50d-f10f-4ff8-b65a-d64b1ebb83c8"),
+			exchange: &microtest.Exchange{
+				Response: microtest.Response{
+					Status: 200,
+					Body:   `{"message": "item deleted"}`,
+				},
+			},
+			uri: "/item/f27ef50d-f10f-4ff8-b65a-d64b1ebb83c8",
+			e:   nil,
+		},
+	}
+
+	s := NewService(Config{})
+	ms := microtest.MockServer(s)
+	defer ms.Server.Close()
+
+	for i, tc := range tests {
+		name := fmt.Sprintf("%d %s", i, tc.name)
+		t.Run(name, func(t *testing.T) {
+			ms.Append(tc.exchange)
+
+			e := s.DeleteItem(tc.uuid)
+
+			// ensure errors match as expected
+			if NotEqualError(e, tc.e) {
+				t.Errorf("expected error %v, got %v", tc.e, e)
+			}
+
+			// ensure the correct uri was called
+			if tc.uri != tc.exchange.Request.RequestURI {
+				t.Errorf("expected uri %s, got %s", tc.uri, tc.exchange.Request.RequestURI)
+			}
+		})
+	}
+}